Almost there... Have to figure out why only one shape gets drawn though.

+++ src/Vertex.java
@@ -1,28 +1,25 @@
-/**
- * Created by flasca50 on 2/7/2016.
- */
 public class Vertex {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Vertex(int x, int y) {
+    public Vertex(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
